deps: upgrade, is building
