+++ packages/palette-docs/package.json
@@ -54,8 +54,8 @@
     "@types/lodash": "4.14.123",
     "@types/react": "17.0.2",
     "@types/styled-components": "4.0.2",
-    "babel-plugin-styled-components": "1.12.0",
-    "gatsby-plugin-typescript": "2.0.13",
+    "babel-plugin-styled-components": "1.13.2",
+    "gatsby-plugin-typescript": "3.11.0",
     "husky": "2.1.0",
     "lint-staged": "8.1.5",
     "netlify-cms-backend-fs": "0.4.4",
@@ -71,41 +71,42 @@
   },
   "dependencies": {
     "@artsy/palette": "^14.51.0",
-    "@babel/core": "^7.2.2",
-    "@mdx-js/mdx": "^0.20.0",
+    "@babel/core": "^7.15.0",
+    "@mdx-js/mdx": "^1.6.22",
+    "@mdx-js/react": "^1.6.22",
     "@mdx-js/tag": "^0.20.0",
     "@reach/router": "^1.2.1",
     "@styled-system/theme-get": "^5.1.2",
-    "babel-preset-gatsby": "^1.10.0",
+    "babel-preset-gatsby": "^1.11.0",
     "d3-ease": "^1.0.5",
-    "docz-utils": "^0.13.6",
-    "gatsby": "2.32.11",
-    "gatsby-image": "^2.0.20",
-    "gatsby-mdx": "^0.3.4",
-    "gatsby-plugin-catch-links": "^2.0.10",
+    "docz-utils": "^2.3.0",
+    "gatsby": "3.11.0",
+    "gatsby-image": "^3.11.0",
+    "gatsby-plugin-catch-links": "^3.11.0",
     "gatsby-plugin-force-trailing-slashes": "^1.0.4",
-    "gatsby-plugin-manifest": "^2.4.30",
-    "gatsby-plugin-offline": "^2.0.16",
-    "gatsby-plugin-react-helmet": "^3.0.5",
+    "gatsby-plugin-manifest": "^3.11.0",
+    "gatsby-plugin-mdx": "^2.11.0",
+    "gatsby-plugin-offline": "^4.11.0",
+    "gatsby-plugin-react-helmet": "^4.11.0",
     "gatsby-plugin-s3": "^0.3.2",
-    "gatsby-plugin-sharp": "^2.6.36",
-    "gatsby-plugin-sitemap": "^2.2.16",
-    "gatsby-plugin-styled-components": "^3.0.4",
-    "gatsby-remark-prismjs": "^3.2.4",
-    "gatsby-source-filesystem": "^2.0.16",
-    "gatsby-transformer-sharp": "^2.5.15",
-    "hast-util-to-string": "^1.0.1",
+    "gatsby-plugin-sharp": "^3.11.0",
+    "gatsby-plugin-sitemap": "^4.7.0",
+    "gatsby-plugin-styled-components": "^4.11.0",
+    "gatsby-remark-prismjs": "^5.8.0",
+    "gatsby-source-filesystem": "^3.11.0",
+    "gatsby-transformer-sharp": "^3.11.0",
+    "hast-util-to-string": "^2.0.0",
     "lodash": "^4.17.11",
-    "marked": "^0.6.0",
-    "prismjs": "^1.15.0",
+    "marked": "^2.1.3",
+    "prismjs": "^1.24.1",
     "prop-types": "^15.6.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-helmet": "^5.2.0",
-    "react-live": "1.12.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-helmet": "^6.1.0",
+    "react-live": "2.3.0",
     "react-powerplug": "^1.0.0",
     "scroll-into-view-if-needed": "^2.2.28",
-    "styled-components": "^4.3.2",
+    "styled-components": "^5.3.0",
     "unstated": "^2.1.1"
   },
   "license": "MIT",
