+++ packages/palette-docs/src/Boot.tsx
@@ -1,10 +1,11 @@
 import { Box, Theme } from "@artsy/palette"
-import { MDXProvider } from "@mdx-js/tag"
+import { MDXProvider } from "@mdx-js/react"
 import { Location } from "@reach/router"
+import { GlobalStyles } from "components/GlobalStyles"
 import { NavState } from "components/Sidebar/NavState"
 import React from "react"
 import { Provider as StateProvider } from "unstated"
-import { GlobalStyles, MarkdownComponents } from "./components/GlobalComponents"
+import { MDXComponents } from "./components/MDXComponents"
 
 export const Boot = ({ element }) => {
   return (
@@ -12,7 +13,7 @@ export const Boot = ({ element }) => {
       {({ location }) => {
         return (
           <StateProvider inject={[new NavState(location.pathname)]}>
-            <MDXProvider components={MarkdownComponents}>
+            <MDXProvider components={MDXComponents}>
               <Theme theme="v3">
                 <Box>
                   <GlobalStyles />
